--- Runtime/Utilities/UX/Utilities.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 17beeb9747ba4bef8df5d27754d7dee5
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
